backend/requirements.txt
Authlib==1.6.2
azure-core==1.35.0
azure-identity==1.24.0
chromadb==1.0.20
Flask==3.1.2
flask-cors==6.0.1
langchain==0.3.27
langchain-chroma==0.2.5
langchain-community==0.3.27
langchain-core==0.3.74
langchain-openai==0.3.31
langchain-text-splitters==0.3.9
langgraph==0.6.6
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.3
langsmith==0.4.16
lxml==6.0.1
Markdown==3.9
markdown-it-py==4.0.0
markdownlit==0.0.7
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.6
mdurl==0.1.2
mmh3==5.2.0
MouseInfo==0.1.3
mpmath==1.3.0
msal==1.33.0
msal-extensions==1.3.1
multidict==6.6.4
mypy_extensions==1.1.0
narwhals==2.1.2
numpy==2.3.2
oauthlib==3.3.1
onnxruntime==1.22.1
openai==1.101.0
packaging==25.0
pdfplumber==0.11.7
pillow==11.3.0
platformdirs==4.4.0
plotly==6.3.0
posthog==5.4.0
prometheus_client==0.22.1
propcache==0.3.2
proto-plus==1.26.1
protobuf==6.31.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyAutoGUI==0.9.54
pybase64==1.4.2
pycparser==2.22
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
pydantic[email]==2.11.7
pydeck==0.9.1
PyGetWindow==0.0.9
Pygments==2.19.2
PyJWT==2.10.1
pymdown-extensions==10.16.1
pymongo==4.14.1
PyMsgBox==1.0.9
PyMuPDF==1.26.3
pyOpenSSL==25.1.0
pyparsing==3.2.3
pypdf==6.0.0
pypdfium2==4.30.0
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
requests==2.32.5
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
s3transfer==0.13.1
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowflake-connector-python==3.17.3
snowflake-snowpark-python==1.38.0
sortedcontainers==2.4.0
soupsieve==2.8
SQLAlchemy==2.0.43
sympy==1.14.0
tenacity==9.1.2
tiktoken==0.11.0
tokenizers==0.21.4
toml==0.10.2
tomlkit==0.13.3
tornado==6.5.2
tqdm==4.67.1
typer==0.16.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
validators==0.35.0
watchfiles==1.1.0
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
gunicorn==21.2.0
google-cloud-storage